+++ src/core/types/common.hpp
@@ -4,7 +4,7 @@
 namespace Ballistic {
     namespace Core {
         namespace Types {
-            typedef double scalar_t;
+            typedef float scalar_t;
             typedef scalar_t v2d_t[2];
             typedef scalar_t v3d_t[3];
         }
